refactor(qrcontainer):remove ucode that doesnt belong in qrContainer
