--- components/folder/FolderContent.js
@@ -1,12 +1,0 @@
-import { View } from 'react-native';
-import React from 'react';
-
-function FolderContent() {
-    return (
-        <View>
-            
-        </View>
-    )
-}
-
-export default FolderContent
